The Goblin Party Update
This update is the first update to lay down the foundations to allow for multiple different types of Goblln models to exist. This means that Goblln is a generalization of the original Code Goblin. Code Goblin still exists. However, Code Goblin is now an an extension to Goblln rather than its own standalone library.